Replace TestCatalogController to CatalogViewController
