+++ FakeNFT/Resources/Assets.xcassets/Contents.json
@@ -2,5 +2,8 @@
   "info" : {
     "author" : "xcode",
     "version" : 1
+  },
+  "properties" : {
+    "generate-swift-asset-symbol-extensions" : "enabled"
   }
 }
